docker push: Improved error msg when pushing a non-existant image.
The prior error message caused confusion.  If a user attempts to push an image up to a registry, but they misspelled (or forgot to properly tag their image) they would see the message 'Repository does not exist', which is not very clear and causes some to think that there might be a problem with the registry or connectivity to it, when the problem was simply just that an image with that tag specified does not exist locally.